requirements.txt
langchain
PyPdf
python-dotenv
streamlit
faiss-cpu
streamlit-extras
openai
tiktoken
